requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.40.0
httpx==0.27.0
pypdf==4.3.1
python-docx==1.1.2
python-multipart==0.0.9

